chore/menu: break into multiple parts
